+++ src/samplers/drawtable.jl
@@ -1,7 +1,7 @@
 # Store an alias table
 immutable DiscreteDistributionTable <: AbstractCategoricalSampler
-	table::Vector{Vector{Int64}}
-	bounds::Vector{Int64}
+	table::Vector{Vector{Int}}
+	bounds::Vector{Int}
 end
 
 # TODO: Test if bit operations can speed up Base64 mod's and fld's
@@ -10,25 +10,25 @@ function DiscreteDistributionTable{T <: Real}(probs::Vector{T})
 	n = length(probs)
 
 	# Convert all Float64's into integers
-	vals = Array(Int64, n)
+	vals = Array(Int, n)
 	for i in 1:n
 		vals[i] = int(probs[i] * 64^9)
 	end
 
 	# Allocate digit table and digit sums as table bounds
-	table = Array(Vector{Int64}, 9)
-	bounds = zeros(Int64, 9)
+	table = Array(Vector{Int}, 9)
+	bounds = zeros(Int, 9)
 
 	# Special case for deterministic distributions
 	for i in 1:n
 		if vals[i] == 64^9
-			table[1] = Array(Int64, 64)
+			table[1] = Array(Int, 64)
 			for j in 1:64
 				table[1][j] = i
 			end
 			bounds[1] = 64^9
 			for j in 2:9
-				table[j] = Array(Int64, 0)
+				table[j] = Array(Int, 0)
 				bounds[j] = 64^9
 			end
 			return DiscreteDistributionTable(table, bounds)
@@ -38,7 +38,7 @@ function DiscreteDistributionTable{T <: Real}(probs::Vector{T})
 	# Fill tables
 	multiplier = 1
 	for index in 9:-1:1
-		counts = Array(Int64, 0)
+		counts = Array(Int, 0)
 		for i in 1:n
 			digit = mod(vals[i], 64)
 			# vals[i] = fld(vals[i], 64)
